backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# LLM providers
openai==1.30.1
google-generativeai==0.5.4
requests==2.32.2          # ollama HTTP calls

# Embedders
cohere==5.5.0

# Databases
PyMySQL==1.1.0
pymongo==4.7.2
DBUtils==3.1.0            # MySQL connection pool

# Vector stores
faiss-cpu==1.8.0
pinecone-client==4.1.0    # optional — only if VECTOR_STORE=pinecone

# Strategy utilities
python-Levenshtein==0.25.1
sqlparse==0.5.0
numpy==1.26.4

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0             # async test client for FastAPI
